Introduced Dependency Injection for database session management in all route handlers. Modified the CRUD Operations to connect and interact with the postgresql database.

+++ models.py
@@ -4,6 +4,6 @@ class Product(BaseModel):
     id: int
     name: str
     description: str
-    price: int
+    price: float
     quantity: int
